requirements.txt
transformers==3.0.2
tokenizers==0.8.0
torchvision==0.5
torch>=1.3
tensorboardX==2.0
numpy
pandas>1.0
tape_proteins==0.4
seaborn==0.11.0
loguru==0.5.3
scikit-learn==0.23.2
seqeval  # dependency for token clf metric calc
